Move and rename GoogleAnalytics service
Pull Request #39 made large changes to how the code is structured. One
of those changes was to put all `service` code in `app/services` and
append `Service` to the name of the class.

This commit makes the GoogleAnalytics service code comply with these
new standards.
